PMS/requirements.txt
algoliasearch-django==4.0.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-admin-honeypot==1.1.0
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
environ==1.0
et_xmlfile==2.0.0
idna==3.10
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
